Transaction module API doc
The document that describes APIs to use for a transaction module (to read and modify account transactions). This is for presentation purposes only (the actual APIs are not configured).
It's written with Swagger (OpenAPI 3.0.0) guidelines.

Another swagger.yaml document has been provided for a user module APIs by the P13 back-end kit (this one has it's APIs operative)